+++ packages/zowe-explorer-api/src/profiles/ProfilesCache.ts
@@ -9,7 +9,7 @@
  *                                                                                 *
  */
 
-import { IProfileLoaded, Logger, CliProfileManager, IProfile, ImperativeConfig } from "@zowe/imperative";
+import * as imperative from "@zowe/imperative";
 import * as path from "path";
 import * as os from "os";
 import { URL } from "url";
@@ -45,14 +45,14 @@ export enum ValidProfileEnum {
 export class ProfilesCache {
     public profilesForValidation: IProfileValidation[] = [];
     public profilesValidationSetting: IValidationSetting[] = [];
-    public allProfiles: IProfileLoaded[] = [];
+    public allProfiles: imperative.IProfileLoaded[] = [];
     protected allTypes: string[];
-    protected profilesByType = new Map<string, IProfileLoaded[]>();
-    protected defaultProfileByType = new Map<string, IProfileLoaded>();
-    protected profileManagerByType = new Map<string, CliProfileManager>();
-    public constructor(protected log: Logger) {}
+    protected profilesByType = new Map<string, imperative.IProfileLoaded[]>();
+    protected defaultProfileByType = new Map<string, imperative.IProfileLoaded>();
+    protected profileManagerByType = new Map<string, imperative.CliProfileManager>();
+    public constructor(protected log: imperative.Logger) {}
 
-    public loadNamedProfile(name: string, type?: string): IProfileLoaded {
+    public loadNamedProfile(name: string, type?: string): imperative.IProfileLoaded {
         for (const profile of this.allProfiles) {
             if (profile.name === name && (type ? profile.type === type : true)) {
                 return profile;
@@ -61,15 +61,15 @@ export class ProfilesCache {
         throw new Error("Could not find profile named: " + name + ".");
     }
 
-    public getDefaultProfile(type = "zosmf"): IProfileLoaded {
+    public getDefaultProfile(type = "zosmf"): imperative.IProfileLoaded {
         return this.defaultProfileByType.get(type);
     }
 
-    public getProfiles(type = "zosmf"): IProfileLoaded[] {
+    public getProfiles(type = "zosmf"): imperative.IProfileLoaded[] {
         return this.profilesByType.get(type);
     }
 
-    public async refresh(apiRegister: ZoweExplorerApi.IApiRegisterClient): Promise<void> {
+    public async refresh(apiRegister?: ZoweExplorerApi.IApiRegisterClient): Promise<void> {
         this.allProfiles = [];
         this.allTypes = [];
         // TODO: Add Base ProfileType in registeredApiTypes
@@ -98,7 +98,7 @@ export class ProfilesCache {
             if (profilesForType && profilesForType.length > 0) {
                 this.allProfiles.push(...profilesForType);
                 this.profilesByType.set(type, profilesForType);
-                let defaultProfile: IProfileLoaded;
+                let defaultProfile: imperative.IProfileLoaded;
                 try {
                     defaultProfile = await profileManager.load({ loadDefault: true });
                 } catch (error) {
@@ -171,8 +171,8 @@ export class ProfilesCache {
         });
     }
 
-    public async directLoad(type: string, name: string): Promise<IProfileLoaded> {
-        let directProfile: IProfileLoaded;
+    public async directLoad(type: string, name: string): Promise<imperative.IProfileLoaded> {
+        let directProfile: imperative.IProfileLoaded;
         const profileManager = this.getCliProfileManager(type);
         if (profileManager) {
             directProfile = await profileManager.load({ name });
@@ -180,11 +180,11 @@ export class ProfilesCache {
         return directProfile;
     }
 
-    public getCliProfileManager(type: string): CliProfileManager {
+    public getCliProfileManager(type: string): imperative.CliProfileManager {
         let profileManager = this.profileManagerByType.get(type);
         if (!profileManager) {
             try {
-                profileManager = new CliProfileManager({
+                profileManager = new imperative.CliProfileManager({
                     profileRootDirectory: path.join(this.getZoweDir(), "profiles"),
                     type,
                 });
@@ -200,8 +200,8 @@ export class ProfilesCache {
         return profileManager;
     }
 
-    public getBaseProfile(): IProfileLoaded {
-        let baseProfile: IProfileLoaded;
+    public getBaseProfile(): imperative.IProfileLoaded {
+        let baseProfile: imperative.IProfileLoaded;
 
         // This functionality will retrieve the saved base profile in the allProfiles array
         for (const baseP of this.allProfiles) {
@@ -212,7 +212,7 @@ export class ProfilesCache {
         return baseProfile;
     }
 
-    protected async deleteProfileOnDisk(ProfileInfo: IProfileLoaded): Promise<void> {
+    protected async deleteProfileOnDisk(ProfileInfo: imperative.IProfileLoaded): Promise<void> {
         await this.getCliProfileManager(ProfileInfo.type).delete({
             profile: ProfileInfo,
             name: ProfileInfo.name,
@@ -224,7 +224,7 @@ export class ProfilesCache {
         profileInfo: Record<string, unknown>,
         profileName: string,
         profileType: string
-    ): Promise<IProfile> {
+    ): Promise<imperative.IProfile> {
         const newProfile = await this.getCliProfileManager(profileType).save({
             profile: profileInfo,
             name: profileName,
@@ -239,10 +239,10 @@ export class ProfilesCache {
      * not initialized it we mock a default value.
      */
     private getZoweDir(): string {
-        ImperativeConfig.instance.loadedConfig = {
+        imperative.ImperativeConfig.instance.loadedConfig = {
             defaultHome: path.join(os.homedir(), ".zowe"),
             envVariablePrefix: "ZOWE",
         };
-        return ImperativeConfig.instance.cliHome;
+        return imperative.ImperativeConfig.instance.cliHome;
     }
 }